Add condition checks and loops for reverse string
String argument is first split and checked if there are more than one
element in the returned array. This way, the string can be revered solo
or if there are more than one element in the array, split each element and
reverse the arrays. Used join to change arrays back into strings.